requirements.txt
discord.py>=2.3
python-dotenv>=1.0
twitchio>=2.9
google-generativeai>=0.3
requests>=2.31
elevenlabs

